moved keybindings method to the controls class
