+++ Devoid/Controls.cs
@@ -1,0 +1,42 @@
+﻿using Microsoft.Xna.Framework.Input;
+using Microsoft.Xna.Framework;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Devoid
+{
+    public class Controls
+    {
+        public Vector2 Keybindings(GameTime gameTime, Player player)
+        {
+            /* Player Movement */
+            var kstate = Keyboard.GetState();
+
+            if (kstate.IsKeyDown(Keys.Up))
+            {
+                // Move the player up
+                player.Position.Y -= player.getSpeed() * (float)gameTime.ElapsedGameTime.TotalSeconds;
+            }
+            else if (kstate.IsKeyDown(Keys.Down))
+            {
+                // Move the player down
+                player.Position.Y += player.getSpeed() * (float)gameTime.ElapsedGameTime.TotalSeconds;
+            }
+            else if (kstate.IsKeyDown(Keys.Left))
+            {
+                // Move the player left
+                player.Position.X -= player.getSpeed() * (float)gameTime.ElapsedGameTime.TotalSeconds;
+            }
+            else if (kstate.IsKeyDown(Keys.Right))
+            {
+                // Move the player right
+                player.Position.X += player.getSpeed() * (float)gameTime.ElapsedGameTime.TotalSeconds;
+            }
+
+            return player.Position;
+        }
+    }
+}
